Css poboljsanja

+++ src/app/components/main-panel/price-list/price-list.ts
@@ -8,19 +8,24 @@ import { CompanyManagmentService } from '../../../services/company.managment.ser
 import { CommonModule } from '@angular/common';
 import { UpsertPriceListModal } from './upsert-price-list-modal/upsert-price-list-modal';
 import { ProductService } from '../../../services/product.service';
+import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
+import { ConfirmDialogBox } from '../../pop-up/confirm-dialog-box/confirm-dialog-box';
+import { take } from 'rxjs';
+import { NotificationService } from '../../../services/notification.service';
 
 @Component({
   selector: 'app-price-list',
-  imports: [CommonModule, ReactiveFormsModule, UpsertPriceListModal],
+  imports: [CommonModule, ReactiveFormsModule],
   templateUrl: './price-list.html',
   styleUrl: './price-list.css'
 })
 export class PriceList {
+
+  modalRef?: BsModalRef<UpsertPriceListModal>;
+  modalRefConfirm?: BsModalRef<ConfirmDialogBox>;
+
   productForm!: FormGroup;
   filterForm: FormGroup;
-
-  showModal = false;
-  isCreateCompanyModal = false;
   modalTitle = '';
 
   companies: Company[] = [];
@@ -29,7 +34,7 @@ export class PriceList {
   role: UserRole;
 
   constructor(private productManagemntService: ProductService, private companyManagmentService: CompanyManagmentService,
-    private authService: AuthService
+    private authService: AuthService, private modalService: BsModalService, private notifyService: NotificationService
   ) {
 
     this.role = this.authService.getUserRole() as UserRole;
@@ -58,7 +63,15 @@ export class PriceList {
         console.log('Products fetched:', this.products);
       },
       error: (error) => {
-        console.error('Error fetching products:', error);
+        if (error.error && error.error.error) {
+          this.notifyService.error(error.error.error);
+        } 
+        else if (typeof error.error === 'string') {
+          this.notifyService.error(error.error);
+        } 
+        else {
+          this.notifyService.error(error);
+        }
       }
     });
 
@@ -67,7 +80,15 @@ export class PriceList {
         this.companies = companies;
       },
       error: (error) => {
-        console.error('Error fetching companies:', error);
+        if (error.error && error.error.error) {
+          this.notifyService.error(error.error.error);
+        } 
+        else if (typeof error.error === 'string') {
+          this.notifyService.error(error.error);
+        } 
+        else {
+          this.notifyService.error(error);
+        }
       }
     });
   }
@@ -83,17 +104,13 @@ export class PriceList {
     });
 
     this.modalTitle = 'Dodaj novi proizvod';
-    this.isCreateCompanyModal = true;
-    this.showModal = true;
+
+    this.callModal(true, true)
 
   }
 
-  closeModal() {
-      this.showModal = false;
-    }
   
   confirmModal(apiResponse: CreateApiResponse<any>) {
-    this.showModal = false;
 
     this.productManagemntService.getAllProducts().subscribe({
 
@@ -101,20 +118,52 @@ export class PriceList {
         this.products = products;
       },
       error: (error) => {
-        console.error('Error fetching products:', error);
+        if (error.error && error.error.error) {
+          this.notifyService.error(error.error.error);
+        } 
+        else if (typeof error.error === 'string') {
+          this.notifyService.error(error.error);
+        } 
+        else {
+          this.notifyService.error(error);
+        }
       }
     });
   }
 
   deleteProduct(productId: number) {
-    this.productManagemntService.deleteProduct(productId).subscribe({
-      next: () => {
-        this.products = this.products.filter(p => p.productId !== productId);
+
+    const modalRefConfirm: BsModalRef = this.modalService.show(ConfirmDialogBox, {
+      initialState: {
+        title: 'Promjena statusa',
+        message: 'Da li ste sigurni da želite izbrisati ovaj proizvod?',
+        onConfirm: () => {
+          this.productManagemntService.deleteProduct(productId).subscribe({
+            next: () => {
+              this.products = this.products.filter(p => p.productId !== productId);
+            },
+            error: (error) => {
+              if (error.error && error.error.error) {
+                this.notifyService.error(error.error.error);
+              } 
+              else if (typeof error.error === 'string') {
+                this.notifyService.error(error.error);
+              } 
+              else {
+                this.notifyService.error(error);
+              }
+            }
+          });
+        },
+        onCancel: () => {
+          this.notifyService.info('Brisanje proizvoda otkazano');
+        } 
       },
-      error: (error) => {
-        console.error('Error deleting product:', error);
-      }
+        ignoreBackdropClick: true,
+        keyboard: false 
     });
+
+
   }
 
   openUpdateOUModal(product: Product) {
@@ -128,10 +177,8 @@ export class PriceList {
     });
 
     this.modalTitle = 'Azuriraj proizvod';
-    this.isCreateCompanyModal = false;
-    this.showModal = true;
 
-
+    this.callModal(true, false);
   }
 
   onFilterSubmit() {
@@ -145,7 +192,15 @@ export class PriceList {
             this.products = products;
           },
           error: (error) => {
-            console.error('Error fetching products:', error);
+            if (error.error && error.error.error) {
+              this.notifyService.error(error.error.error);
+            } 
+            else if (typeof error.error === 'string') {
+              this.notifyService.error(error.error);
+            } 
+            else {
+              this.notifyService.error(error);
+            }
           }
       });
     }
@@ -156,9 +211,38 @@ export class PriceList {
           this.products = products;
         },
         error: (error) => {
-          console.error('Error fetching products:', error);
+          if (error.error && error.error.error) {
+            this.notifyService.error(error.error.error);
+          } 
+          else if (typeof error.error === 'string') {
+            this.notifyService.error(error.error);
+          } 
+          else {
+            this.notifyService.error(error);
+          }
         }
       });
     }
+  }
+
+
+  callModal(show: boolean, isCreate: boolean) {
+      
+    this.modalRef = this.modalService.show(UpsertPriceListModal, {
+      class: 'modal-dialog modal-xl modal-fullscreen-md-down modal-dialog-centered',
+      initialState: {
+        show: show,
+        isCreateCompanyModal: isCreate,
+        productForm: this.productForm,
+        companies: this.companies,
+        title: this.modalTitle
+      },
+      ignoreBackdropClick: true,
+      keyboard: false 
+    });
+
+    this.modalRef.onHidden!.pipe(take(1)).subscribe(() => {
+      this.onFilterSubmit();
+    });
   }
 }
